Move set_vtkwriter_mode to fileio.py
